change the type of wake_time(unsignedint->int64_t)

+++ include/threads/thread.h
@@ -107,7 +107,7 @@ struct thread {
 	/* Owned by thread.c. */
 	struct intr_frame tf;               /* Information for switching */
 	unsigned magic;                     /* Detects stack overflow. */
-	unsigned int wake_time;             /* the time is thread wake up*/
+	int64_t wake_time;             /* the time is thread wake up*/
 };
 
 /* If false (default), use round-robin scheduler.
